+++ packages/devtools-ui/src/index.ts
@@ -8,3 +8,12 @@ export { Tag } from './components/tag'
 export type { JsonTreeWebComponentProps } from './components/tree'
 export type { ButtonProps } from './components/button'
 export type { CustomElements } from './types'
+export { MainPanel } from './components/main-panel'
+export {
+  Section,
+  SectionTitle,
+  SectionDescription,
+  SectionIcon,
+} from './components/section'
+export { Header, HeaderLogo } from './components/header'
+export { useTheme, ThemeContextProvider } from './components/theme'