+++ packages/devtools/src/components/main-panel.tsx
@@ -1,7 +1,10 @@
 import clsx from 'clsx'
+import { DrawClientProvider } from '../context/draw-context'
 import { useDevtoolsSettings, useHeight } from '../context/use-devtools-context'
 import { useStyles } from '../styles/use-styles'
 import { TANSTACK_DEVTOOLS } from '../utils/storage'
+import { usePiPWindow } from '../context/pip-context'
+
 import type { Accessor, JSX } from 'solid-js'
 
 export const MainPanel = (props: {
@@ -12,20 +15,26 @@ export const MainPanel = (props: {
   const styles = useStyles()
   const { height } = useHeight()
   const { settings } = useDevtoolsSettings()
+  const pip = usePiPWindow()
   return (
     <div
       id={TANSTACK_DEVTOOLS}
       style={{
-        height: height() + 'px',
+        height: pip().pipWindow ? '100vh' : height() + 'px',
       }}
       class={clsx(
-        styles().devtoolsPanelContainer(settings().panelLocation),
+        styles().devtoolsPanelContainer(
+          settings().panelLocation,
+          Boolean(pip().pipWindow),
+        ),
         styles().devtoolsPanelContainerAnimation(props.isOpen(), height()),
         styles().devtoolsPanelContainerVisibility(props.isOpen()),
         styles().devtoolsPanelContainerResizing(props.isResizing),
       )}
     >
-      {props.children}
+      <DrawClientProvider animationMs={400}>
+        {props.children}
+      </DrawClientProvider>
     </div>
   )
 }