+++ packages/devtools/src/components/tabs.tsx
@@ -1,8 +1,11 @@
 import clsx from 'clsx'
 import { For } from 'solid-js'
+import { PiP, X } from '@tanstack/devtools-ui/icons'
 import { useStyles } from '../styles/use-styles'
 import { useDevtoolsState } from '../context/use-devtools-context'
+import { useDrawContext } from '../context/draw-context'
 import { tabs } from '../tabs'
+import { usePiPWindow } from '../context/pip-context'
 
 interface TabsProps {
   toggleOpen: () => void
@@ -11,6 +14,13 @@ interface TabsProps {
 export const Tabs = (props: TabsProps) => {
   const styles = useStyles()
   const { state, setState } = useDevtoolsState()
+  const pipWindow = usePiPWindow()
+  const handleDetachment = () => {
+    pipWindow().requestPipWindow(
+      `width=${window.innerWidth},height=${state().height},top=${window.screen.height},left=${window.screenLeft}}`,
+    )
+  }
+  const { hoverUtils } = useDrawContext()
 
   return (
     <div class={styles().tabContainer}>
@@ -20,31 +30,39 @@ export const Tabs = (props: TabsProps) => {
             type="button"
             onClick={() => setState({ activeTab: tab.id })}
             class={clsx(styles().tab, { active: state().activeTab === tab.id })}
+            onMouseEnter={() => {
+              if (tab.id === 'plugins') hoverUtils.enter()
+            }}
+            onMouseLeave={() => {
+              if (tab.id === 'plugins') hoverUtils.leave()
+            }}
           >
-            {tab.icon}
+            {tab.icon()}
           </button>
         )}
       </For>
-      <button
-        type="button"
-        class={clsx(styles().tab, 'close')}
-        onClick={() => props.toggleOpen()}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          stroke-width="2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
+      {pipWindow().pipWindow !== null ? null : (
+        <div
+          style={{
+            'margin-top': 'auto',
+          }}
         >
-          <path d="M18 6 6 18" />
-          <path d="m6 6 12 12" />
-        </svg>
-      </button>
+          <button
+            type="button"
+            class={clsx(styles().tab, 'detach')}
+            onClick={handleDetachment}
+          >
+            <PiP />
+          </button>
+          <button
+            type="button"
+            class={clsx(styles().tab, 'close')}
+            onClick={() => props.toggleOpen()}
+          >
+            <X />
+          </button>
+        </div>
+      )}
     </div>
   )
 }