+++ packages/devtools/src/context/use-devtools-context.ts
@@ -1,6 +1,7 @@
-import { createMemo, useContext } from 'solid-js'
+import { createEffect, createMemo, useContext } from 'solid-js'
 import { DevtoolsContext } from './devtools-context.jsx'
-/* import type { DevtoolsPlugin } from './devtools-context' */
+import { useDrawContext } from './draw-context.jsx'
+
 import type { DevtoolsStore } from './devtools-store.js'
 
 /**
@@ -17,11 +18,30 @@ const useDevtoolsContext = () => {
   return context
 }
 
+export function useTheme() {
+  const { settings, setSettings } = useDevtoolsSettings()
+  const theme = createMemo(() => settings().theme)
+  return {
+    theme,
+    setTheme: (theme: DevtoolsStore['settings']['theme']) =>
+      setSettings({ theme }),
+  }
+}
+
 export const usePlugins = () => {
   const { store, setStore } = useDevtoolsContext()
+  const { setForceExpand } = useDrawContext()
 
   const plugins = createMemo(() => store.plugins)
   const activePlugin = createMemo(() => store.state.activePlugin)
+
+  createEffect(() => {
+    if (activePlugin() == null) {
+      setForceExpand(false)
+    } else {
+      setForceExpand(true)
+    }
+  })
 
   const setActivePlugin = (pluginId: string) => {
     setStore((prev) => ({