+++ packages/devtools/src/core.tsx
@@ -7,6 +7,7 @@ import {
 } from '@tanstack/devtools-ui'
 import { DevtoolsProvider } from './context/devtools-context'
 import { initialState } from './context/devtools-store'
+import { PiPProvider } from './context/pip-context'
 import type { ClientEventBusConfig } from '@tanstack/devtools-event-bus/client'
 import type {
   TanStackDevtoolsConfig,
@@ -64,6 +65,9 @@ export class TanStackDevtoolsCore {
   }
 
   mount<T extends HTMLElement>(el: T) {
+    //  tsup-preset-solid statically replaces this variable during build, which eliminates this code from server bundle
+    if (import.meta.env.SSR) return
+
     if (this.#isMounted) {
       throw new Error('Devtools is already mounted')
     }
@@ -77,9 +81,11 @@ export class TanStackDevtoolsCore {
       registerButtonComponent()
       return (
         <DevtoolsProvider plugins={this.#plugins} config={this.#config}>
-          <Portal mount={mountTo}>
-            <Devtools />
-          </Portal>
+          <PiPProvider>
+            <Portal mount={mountTo}>
+              <Devtools />
+            </Portal>
+          </PiPProvider>
         </DevtoolsProvider>
       )
     }, mountTo)