+++ packages/event-bus/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tanstack/devtools-event-bus
 
+## 0.3.2
+
+### Patch Changes
+
+- fix issue with broadcast channel not emitting functions properly and failing ([#106](https://github.com/TanStack/devtools/pull/106))
+
+## 0.3.1
+
+### Patch Changes
+
+- fix a bug for server event bus not connecting with clients properly ([#88](https://github.com/TanStack/devtools/pull/88))
+
+## 0.3.0
+
+### Minor Changes
+
+- removed CJS support, added detached mode to devtools ([#70](https://github.com/TanStack/devtools/pull/70))
+
 ## 0.2.2
 
 ### Patch Changes